refactor: Separate Verifiable Credential JWT logic
closes #446
